requirements.txt
absl-py==1.3.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.12.0
bleach==5.0.1
blinker==1.6.2
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
ConfigArgParse==1.5.3
contourpy==1.0.6
cupy==11.4.0
cycler==0.11.0
dearpygui==1.8.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
DracoPy==1.3.0
einops==0.3.2
entrypoints==0.4
executing==1.2.0
facenet-pytorch==2.5.3
fastjsonschema==2.16.2
fastrlock==0.8.1
Flask==2.3.2
fonttools==4.38.0
fqdn==1.5.1
freetype-py==2.3.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
fvcore==0.1.5.post20221221
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
icecream==2.1.3
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==6.0.0
importlib-resources==5.10.2
iopath==0.1.10
ipdb==0.13.11
ipykernel==6.19.4
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jittor==1.3.6.10
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==7.4.8
jupyter-console==6.4.4
jupyter_core==5.1.2
jupyter-events==0.5.0
jupyter_packaging==0.12.3
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.18.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kornia==0.4.1
lightning-utilities==0.8.0
llvmlite==0.35.0
lpips==0.1.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nerfacc==0.3.3
nest-asyncio==1.5.6
networkx==2.8.8
ninja==1.11.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.52.0
numpy==1.23.5
nvdiffrast==0.3.0
oauthlib==3.2.2
omegaconf==2.2.3
open3d==0.13.0
opencv-python==4.5.3.56
openmesh==1.2.1
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
plyfile==0.7.2
pooch==1.7.0
portalocker==2.7.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.4
pycparser==2.21
pyDeprecate==0.3.1
pyglet==1.5.0
Pygments==2.14.0
pyhocon==0.3.59
PyMCubes==0.1.2
pymeshlab==2022.2.post3
PyOpenGL==3.1.0
pyparsing==2.4.7
pyquaternion==0.9.9
pyransac3d==0.6.0
pyrender==0.1.45
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytorch-lightning==1.9.1
pytorch-ranger==0.1.1
pytz==2022.7
pyvista==0.38.5
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.0.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
scooby==0.7.1
Send2Trash==1.8.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.2.0
terminado==0.17.1
test-tube==0.7.5
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
tinycudann==1.7
tomli==2.0.1
tomlkit==0.11.6
torch==1.12.1+cu102
torch-efficient-distloss==0.1.3
torch-ema==0.3
torch-optimizer==0.3.0
torchmetrics==0.11.0
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
trimesh==3.17.1
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.13
vtk==9.2.6
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.3.6
wheel==0.37.1
widgetsnbextension==4.0.5
yacs==0.1.8
yarl==1.8.2
zipp==3.11.0
